Add function to replace all blank spans with corresponding user input
