+++ constants.js
@@ -1,33 +1,6 @@
-const zooParagraph = document.getElementById('zoo');
-const inputDisplay = document.getElementById('inputs');
+// Constants not yet used:
 
-
-const aOrAn = document.getElementById('a-or-an');
-
-const adj1 = document.getElementById('adj-1');
-const noun1 = document.getElementById('noun-1');
-const verbPast1 = document.getElementById('verb-pt-1');
-const adverb1 = document.getElementById('adverb-1');
-const adj2 = document.getElementById('adj-2');
-const noun2 = document.getElementById('noun2');
-const noun3 = document.getElementById('noun-3');
-const adj3 = document.getElementById('adj-3');
-const verb1 = document.getElementById('verb-1');
-const adverb2 = document.getElementById('adverb-2');
-const verbPast2 = document.getElementById('verb-pt-2');
-const adj4 = document.getElementById('adj-4');
-
-const adj1Input = document.getElementById('adj-1-input');
-const noun1Input = document.getElementById('noun-1-input');
-const verbPast1Input = document.getElementById('verb-pt-1-input');
-const adverb1Input = document.getElementById('adverb-1-input');
-const adj2Input = document.getElementById('adj-2-input');
-const noun2Input = document.getElementById('noun-2-input');
-const noun3Input = document.getElementById('noun-3-input');
-const adj3Input = document.getElementById('adj-3-input');
-const verb1Input = document.getElementById('verb-1-input');
-const adverb2Input = document.getElementById('adverb-2-input');
-const verbPast2Input = document.getElementById('verb-pt-2-input');
-const adj4Input = document.getElementById('adj-4-input');
-
-const submitButton = document.getElementById('submitBtn');
+// instances where 'a' or 'an' must be determined
+const aOrAn1 = document.getElementById('a-or-an-1');
+const aOrAn2 = document.getElementById('a-or-an-2');
+const aOrAn3 = document.getElementById('a-or-an-3');
