+++ index.html
@@ -22,20 +22,48 @@
     <main>
 
       <div id="zoo">
-        Today I went to the zoo. I saw <span id="a-or-an">a(n)</span>
-        <span id="adj-1">(adjective)</span>
-        <span id="noun-1">(Noun)</span> jumping up and down in its tree. He
-        <span id="verb-pt-1">(verb, past tense)</span>
-        <span id="adverb-1">(adverb)</span> through the large tunnel that led to
-        its <span id="adj-2">(adjective)</span> <span id="noun-2">(noun)</span>.
-        I got some peanuts and passed them through the cage to a gigantic gray
-        <span id="noun-3">(noun)</span> towering above my head. Feeding that
-        animal made me hungry. I went to get a
-        <span id="adj-3">(adjective)</span> scoop of ice cream. It filled my
-        stomach. Afterwards I had to<span id="verb-1">(verb)</span>
-        <span id="adverb-2">(adverb)</span> to catch our bus. When I got home I
-        <span id="verb-pt-2">(verb, past tense)</span> my mom for a
-        <span id="adj-4">(adjective)</span> day at the zoo.
+        Today I went to the zoo. I saw 
+        
+        <span id="a-or-an-1">a(n)</span> 
+        <span id="adj-1">(adjective)</span> 
+        <span id="noun-1">(Noun)</span>
+         
+         jumping up and down in its tree. He 
+        
+        <span id="verb-pt-1">(verb, past tense)</span> 
+        <span id="adverb-1">(adverb)</span>
+        
+         through the large tunnel that led to its 
+        
+        <span id="adj-2">(adjective)</span> 
+        <span id="noun-2">(noun)</span>.
+
+        I got some peanuts and passed them through the cage to a gigantic gray 
+        
+        <span id="noun-3">(noun)</span> 
+        
+        towering above my head. Feeding that
+        animal made me hungry. I went to get 
+        
+        <span id="a-or-an-2">a(n)</span>  
+        
+        <span id="adj-3">(adjective)</span> 
+        
+        scoop of ice cream. It filled my stomach. Afterwards I had to 
+        <span id="verb-1">(verb)</span> 
+        <span id="adverb-2">(adverb)</span>
+        
+         to catch our bus. When I got home I 
+        
+        <span id="verb-pt-2">(verb, past tense)</span> 
+        
+        my mom for 
+        
+        <span id="a-or-an-3">a(n)</span>  
+        
+        <span id="adj-4">(adjective)</span> 
+        
+        day at the zoo.
       </div>
 
       <div id="inputs">
@@ -47,7 +75,7 @@
         <input id="noun-2-input" placeholder="noun">
         <input id="noun-3-input" placeholder="noun">
         <input id="adj-3-input" placeholder="adjective">
-        <input id="verb-1-input" placeholder="verb: present-tense">
+        <input id="verb-1-input" placeholder="verb: present tense">
         <input id="adverb-2-input" placeholder="adverb">
         <input id="verb-pt-2-input" placeholder="verb: past tense">
         <input id="adj-4-input" placeholder="adjective">
